moved ModuleParametersExt to src/module/module.rs
